numba func manager
