+++ tests/warp_drive/pycuda/test_action_sampler.py
@@ -10,7 +10,7 @@ import numpy as np
 import torch
 
 from warp_drive.managers.pycuda.pycuda_data_manager import PyCUDADataManager
-from warp_drive.managers.function_manager import CUDAFunctionManager, CUDASampler
+from warp_drive.managers.pycuda.pycuda_function_manager import PyCUDAFunctionManager, PyCUDASampler
 from warp_drive.utils.common import get_project_root
 from warp_drive.utils.constants import Constants
 from warp_drive.utils.data_feed import DataFeed
@@ -29,12 +29,12 @@ class TestActionSampler(unittest.TestCase):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.dm = PyCUDADataManager(num_agents=5, episode_length=1, num_envs=2)
-        self.fm = CUDAFunctionManager(
+        self.fm = PyCUDAFunctionManager(
             num_agents=int(self.dm.meta_info("n_agents")),
             num_envs=int(self.dm.meta_info("n_envs")),
         )
         self.fm.load_cuda_from_binary_file(f"{_CUBIN_FILEPATH}/test_build.fatbin")
-        self.sampler = CUDASampler(function_manager=self.fm)
+        self.sampler = PyCUDASampler(function_manager=self.fm)
         self.sampler.init_random(seed=None)
 
     def test_agent_action_distribution(self):
