+++ tests/warp_drive/pycuda/test_env_reset.py
@@ -10,9 +10,9 @@ import numpy as np
 import torch
 
 from warp_drive.managers.pycuda.pycuda_data_manager import PyCUDADataManager
-from warp_drive.managers.function_manager import (
-    CUDAEnvironmentReset,
-    CUDAFunctionManager,
+from warp_drive.managers.pycuda.pycuda_function_manager import (
+    PyCUDAEnvironmentReset,
+    PyCUDAFunctionManager,
 )
 from warp_drive.utils.common import get_project_root
 from warp_drive.utils.data_feed import DataFeed
@@ -30,12 +30,12 @@ class TestEnvironmentReset(unittest.TestCase):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.dm = PyCUDADataManager(num_agents=5, num_envs=2, episode_length=2)
-        self.fm = CUDAFunctionManager(
+        self.fm = PyCUDAFunctionManager(
             num_agents=int(self.dm.meta_info("n_agents")),
             num_envs=int(self.dm.meta_info("n_envs")),
         )
         self.fm.load_cuda_from_binary_file(f"{_CUBIN_FILEPATH}/test_build.fatbin")
-        self.resetter = CUDAEnvironmentReset(function_manager=self.fm)
+        self.resetter = PyCUDAEnvironmentReset(function_manager=self.fm)
 
     def test_reset_for_different_dim(self):
