+++ tests/warp_drive/pycuda/test_function_manager.py
@@ -10,10 +10,10 @@ import numpy as np
 import torch
 
 from warp_drive.managers.pycuda.pycuda_data_manager import PyCUDADataManager
-from warp_drive.managers.function_manager import (
-    CUDAEnvironmentReset,
-    CUDAFunctionManager,
-    CUDALogController,
+from warp_drive.managers.pycuda.pycuda_function_manager import (
+    PyCUDAEnvironmentReset,
+    PyCUDAFunctionManager,
+    PyCUDALogController,
 )
 from warp_drive.utils.common import get_project_root
 from warp_drive.utils.constants import Constants
@@ -26,9 +26,9 @@ _ACTIONS = Constants.ACTIONS
 
 
 def cuda_dummy_step(
-    function_manager: CUDAFunctionManager,
+    function_manager: PyCUDAFunctionManager,
     data_manager: PyCUDADataManager,
-    env_resetter: CUDAEnvironmentReset,
+    env_resetter: PyCUDAEnvironmentReset,
     target: int,
     step: int,
 ):
@@ -61,13 +61,13 @@ class TestFunctionManager(unittest.TestCase):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.dm = PyCUDADataManager(num_agents=5, num_envs=2, episode_length=2)
-        self.fm = CUDAFunctionManager(
+        self.fm = PyCUDAFunctionManager(
             num_agents=int(self.dm.meta_info("n_agents")),
             num_envs=int(self.dm.meta_info("n_envs")),
         )
         self.fm.load_cuda_from_binary_file(f"{_CUBIN_FILEPATH}/test_build.fatbin")
-        self.dc = CUDALogController(function_manager=self.fm)
-        self.resetter = CUDAEnvironmentReset(function_manager=self.fm)
+        self.dc = PyCUDALogController(function_manager=self.fm)
+        self.resetter = PyCUDAEnvironmentReset(function_manager=self.fm)
 
     def test_step(self):
         self.fm.initialize_functions(["testkernel"])
