+++ warp_drive/numba_includes/core/__init__.py
@@ -4,8 +4,3 @@
 # For full license text, see the LICENSE file in the repo root
 # or https://opensource.org/licenses/BSD-3-Clause
 
-from pathlib import Path
-
-
-def get_project_root() -> Path:
-    return Path(__file__).absolute().parent.parent.parent
