+++ warp_drive/numba_includes/template_env_runner.txt
@@ -1,0 +1,11 @@
+// Copyright (c) 2021, salesforce.com, inc.
+// All rights reserved.
+// SPDX-License-Identifier: BSD-3-Clause
+// For full license text, see the LICENSE file in the repo root
+// or https://opensource.org/licenses/BSD-3-Clause
+
+from warpdrive.numba_includes.env_config import *
+from warpdrive.numba_includes.core.random import *
+from warpdrive.numba_includes.core.reset import *
+
+from <<ENV_NUMBA>> import *
