+++ warp_drive/numba_includes/test_build.py
@@ -1,0 +1,10 @@
+from warpdrive.numba_includes.env_config import *
+from warpdrive.numba_includes.core.random import *
+from warpdrive.numba_includes.core.reset import *
+
+from example_envs.dummy_env.test_step import *
+from example_envs.tag_continuous.tag_continuous_step_numba import *
+
+wkNumberEnvs = 2
+wkNumberAgents = 5
+wkBlocksPerEnv = 1
