+++ warp_drive/utils/pycuda/__init__.py
@@ -3,9 +3,3 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # For full license text, see the LICENSE file in the repo root
 # or https://opensource.org/licenses/BSD-3-Clause
-
-from pathlib import Path
-
-
-def get_project_root() -> Path:
-    return Path(__file__).absolute().parent.parent.parent
